refactor: create regular board folder structure
