+++ frontend/src/components/CreateBoard/RegularBoard/ParticipantsTab/ListParticipants/ListParticipants.tsx
