--- frontend/src/components/CreateBoard/RegularBoard/ParticipantsTab/RadioGroupParticipants/styles.tsx
@@ -1,8 +1,0 @@
-import { styled } from '@stitches/react';
-
-const FormStyled = styled('form', {
-  width: '100%',
-  '@media (min-width: 1600px)': { width: '70%' },
-});
-
-export { FormStyled };
