+++ frontend/src/components/CreateBoard/RegularBoard/ParticipantsTab/SelectTeam/SelectTeam.tsx
@@ -1,32 +1,30 @@
-import React, { useCallback, useEffect, useMemo } from 'react';
-import Icon from '@/components/Primitives/Icons/Icon/Icon';
-import Text from '@/components/Primitives/Text/Text';
-import { teamsOfUser } from '@/store/team/atom/team.atom';
-import { useRecoilState, useRecoilValue } from 'recoil';
-import { createBoardTeam } from '@/store/createBoard/atoms/create-board.atom';
-import { TeamUserRoles } from '@/utils/enums/team.user.roles';
-import Flex from '@/components/Primitives/Layout/Flex/Flex';
-import { BoardUserRoles } from '@/utils/enums/board.user.roles';
-import useCreateBoard from '@/hooks/useCreateBoard';
-import { useRouter } from 'next/router';
 import { useSession } from 'next-auth/react';
+import { useRouter } from 'next/router';
+import React, { useCallback, useEffect, useMemo } from 'react';
+import { useRecoilState, useRecoilValue } from 'recoil';
+
+import Icon from '@/components/Primitives/Icons/Icon/Icon';
 import {
   Select,
+  SelectContent,
+  SelectIcon,
   SelectTrigger,
   SelectValue,
-  SelectIcon,
-  SelectContent,
 } from '@/components/Primitives/Inputs/Select/Select';
-import { HelperTextWrapper } from './styles';
+import Flex from '@/components/Primitives/Layout/Flex/Flex';
+import Text from '@/components/Primitives/Text/Text';
+import useCreateBoard from '@/hooks/useCreateBoard';
+import { createBoardTeam } from '@/store/createBoard/atoms/create-board.atom';
+import { teamsOfUser } from '@/store/team/atom/team.atom';
+import { BoardUserRoles } from '@/utils/enums/board.user.roles';
+import { TeamUserRoles } from '@/utils/enums/team.user.roles';
 
 const SelectTeam = () => {
   const router = useRouter();
   const routerTeam = router.query.team as string;
   const { data: session } = useSession();
 
-  /**
-   * Recoil Atoms and Hooks
-   */
+  // Recoil Atoms and Hooks
   const [selectedTeam, setSelectedTeam] = useRecoilState(createBoardTeam);
   const teams = useRecoilValue(teamsOfUser);
   const { setCreateBoardData, teamMembers } = useCreateBoard(selectedTeam);
@@ -94,7 +92,7 @@ const SelectTeam = () => {
   }, [routerTeam, createBoard, selectedTeam]);
 
   return (
-    <Flex direction="column" css={{ width: '100%' }}>
+    <Flex direction="column" css={{ flex: 1 }}>
       <Select
         disabled={numberOfTeams === 0}
         defaultValue={teamsNames.find((option) => option.value === selectedTeam?.id)?.value}
@@ -110,22 +108,19 @@ const SelectTeam = () => {
             </Text>
             <SelectValue />
           </Flex>
-          <SelectIcon className="SelectIcon" asChild>
+          <SelectIcon className="SelectIcon">
             <Icon name="arrow-down" />
           </SelectIcon>
         </SelectTrigger>
         <SelectContent options={teamsNames} />
       </Select>
       {numberOfTeams === 0 && (
-        <Flex justify="start">
-          <HelperTextWrapper css={{ mt: '$8' }} gap="4">
-            <Icon css={{ width: '$24', height: '$24' }} name="info" />
-
-            <Text hint color="dangerBase">
-              In order to create a team board, you must be team-admin or stakeholder of at least one
-              team.
-            </Text>
-          </HelperTextWrapper>
+        <Flex justify="start" align="center" gap="4" css={{ mt: '$8', color: '$dangerBase' }}>
+          <Icon size={16} name="info" />
+          <Text hint color="dangerBase">
+            In order to create a team board, you must be team-admin or stakeholder of at least one
+            team.
+          </Text>
         </Flex>
       )}
     </Flex>
